requirements.txt
asgiref==3.2.3
coverage==5.0.3
Django==3.0.3
django-debug-toolbar==2.2
djangorestframework==3.11.0
isort==4.3.21
pytz==2019.3
sqlparse==0.3.0
